requirements.txt
selenium==3.141.0
Appium_Python_Client==1.0.2
